+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version 2.2.1 *(In Development)*
 
  * New `setTypeface(Typeface)` and `getTypeface()` methods for title indicator.
    (Thanks Dimitri Fedorov)
+ * Added styled tab indicator sample.
 
 
 Version 2.2.0 *(2012-12-13)*